% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centromeres_hg19.R
\docType{data}
\name{centromeres_hg19}
\alias{centromeres_hg19}
\title{Centromeric regions for genome build hg19.}
\format{
A data with 24 rows and 3 columns:
\describe{
  \item{chr}{chromosome}
  \item{cstart}{centromere starting genome location}
  \item{cend}{centromere end genome location}
}
}
\usage{
centromeres_hg19
}
\description{
A table containg information on centromeric regions for each chromosomes from genome build hg19 extracted from Table Browser.
Group - all tables, Table - gap, Filter type - centromere.
}
\keyword{datasets}
