% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rf_models.R
\name{create_rf_models}
\alias{create_rf_models}
\title{Create model for CNA estimation}
\usage{
create_rf_models(train_data)
}
\arguments{
\item{train_data}{from get_train_dat function}
}
\description{
Create random forest model for diploid level estimation and alteration estimation
}
