% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSNP_GRCh38.R
\docType{data}
\name{dbSNP_GRCh38}
\alias{dbSNP_GRCh38}
\title{Single nucleotide polymorphisms from dbSNP database.}
\format{
Character vector of length 15160915
}
\usage{
dbSNP_GRCh38
}
\description{
A vector containing single nucleotide polymorphisms included in the dbSNP database, version GCF_000001405.38.
Used for filtering only know SNPs.
}
\keyword{datasets}
