% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSNP_hg19.R
\docType{data}
\name{dbSNP_hg19}
\alias{dbSNP_hg19}
\title{Single nucleotide polymorphisms from dbSNP database.}
\format{
Character vector of length 12210680
}
\usage{
dbSNP_hg19
}
\description{
A vector containing single nucleotide polymorphisms included in the dbSNP database corresponding to GRCh37.75
Used for filtering only know SNPs.
}
\keyword{datasets}
