% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diploid_standard_hg19.R
\docType{data}
\name{diploid_standard_hg19}
\alias{diploid_standard_hg19}
\title{Diploid standard hg19.}
\format{
A data with 57779 rows (genes) and 40 columns:
\describe{
  \item{ENSG}{ensembl gene id}
}
}
\usage{
diploid_standard_hg19
}
\description{
A table containg count information from 40 samples without any large CNVs (confirmed by SNP array), which are used to
center the transcription data and calculate the per gene log2 fold change.
}
\keyword{datasets}
