% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diploid_standard_hg38.R
\docType{data}
\name{diploid_standard_hg38}
\alias{diploid_standard_hg38}
\title{Diploid standard hg38.}
\format{
A data with 60622 rows (genes) and 45 columns:
\describe{
  \item{ENSG}{ensembl gene id}
}
}
\usage{
diploid_standard_hg38
}
\description{
A table containg count information from 44 samples without any large CNVs, which are used to
center the transcription data and calculate the per gene log2 fold change.
}
\keyword{datasets}
