% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annot_hg38.R
\docType{data}
\name{gene_annot_hg38}
\alias{gene_annot_hg38}
\title{Gene annotation build hg38.}
\format{
A data with 60616 rows and 4 columns:
\describe{
  \item{chr}{chromosome location}
  \item{start}{starting genome location}
  \item{end}{end genome location}
  \item{ENSG}{ensembl gene id}
}
}
\usage{
gene_annot_hg38
}
\description{
A table containg information on chromosomes, start and end locations of ensembl gene ids.
The data was obtained from genome build hg38
}
\keyword{datasets}
