% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{get_norm_exp}
\alias{get_norm_exp}
\title{Importing packages}
\usage{
get_norm_exp(
  sample_table,
  sample_num,
  standard_samples,
  minReadCnt,
  samp_prop,
  weight_table,
  weight_samp_prop,
  batch,
  generate_weights
)
}
\description{
Importing packages
}
