% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_train_dat.R
\name{get_train_dat}
\alias{get_train_dat}
\title{Model data preparation.}
\usage{
get_train_dat(config, metadata, weight_table, standard)
}
\description{
This function prepares the expression and SNV data for modelling the CNA
}
