% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoautosomal_regions_hg38.R
\docType{data}
\name{pseudoautosomal_regions_hg38}
\alias{pseudoautosomal_regions_hg38}
\title{Pseudoautosomal regions, genome GRCh38.}
\format{
A data table with 4 rows and 4 columns:
\describe{
  \item{parstart}{starting location of a pseudoautosomal region}
  \item{parend}{end location of a pseudoauuildtosomal region}
  \item{chr}{chromosome}
  \item{reg}{pseudoautosomal region}
}
 @source \url{https://www.ncbi.nlm.nih.gov/grc/human/}
}
\usage{
pseudoautosomal_regions_hg38
}
\description{
A table with pseudoautosomal information, build GRCh 38
}
\keyword{datasets}
