% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqCNV_wrappper.R
\name{RNAseqCNV_wrapper}
\alias{RNAseqCNV_wrapper}
\title{RNAseqCNV_wrapper}
\usage{
RNAseqCNV_wrapper(
  config,
  metadata,
  snv_format = "vcf",
  adjust = TRUE,
  arm_lvl = TRUE,
  estimate_lab = TRUE,
  referData = refDataExp,
  keptSNP = keepSNP,
  par_region = par_reg,
  centr_refer = centr_ref,
  weight_tab = weight_table,
  generate_weights = FALSE,
  model_gend = model_gender,
  model_dip = model_dipl,
  model_alter = model_alt,
  model_alter_noSNV = model_noSNV,
  chroms = chrs,
  batch = FALSE,
  dipl_standard = diploid_standard[, c(1:20, 41)],
  scale_cols = scaleCols,
  dpRatioChromEdge = dpRatioChrEdge,
  minDepth = 20,
  minReadCnt = 3,
  samp_prop = 0.8,
  weight_samp_prop = 1
)
}
\arguments{
\item{config}{R script assigning paths to needed directories into variables: 1. count_dir - path to a directory with count files, 2. snv_dir - path to a directory with files with snv information
(either vcf or custom tabular data), out_dir - path to an output directory. More detailed description can be found in the package README file.}

\item{metadata}{path to a metadata table with three columns. First colum: sample names, second column: file names of count files, third column: file names of snv files. There should be no header.
More information is included in the package README file.}

\item{adjust}{logical value, If TRUE, expression is centered according to the random forest estimated diploid chromosomes. Default = TRUE.}

\item{arm_lvl}{logical value, If TRUE, arm_lvl figures will be printed (increases run-time significantly). Defaul = TRUE.}

\item{estimate_lab}{logical value, If TRUE, CNV estimation labels will be included in the final figure.}

\item{referData}{table, reference data for gene annotation with ensamble ids}

\item{keptSNP}{vector of realiable SNPs to keep for the MAF graphs}

\item{par_region}{table with pseudoautosomal regions. These regions will be filtered out.}

\item{centr_refer}{table with chromosomal centromeric locations.}

\item{weight_tab}{table with per-gene weight for calculating weighted quantiles for the boxplots in the main figure.}

\item{generate_weights}{logical value, if TRUE, weights for calculating weighted quantiles will be contructed from variance and depth of the analyzed cohort of samples. If batch is TRUE, the weights will be analyzed
from the batch of input samples, if FALSE the weight will be generate from joined diploid standard and analyzed sample.}

\item{model_gend}{random forest model for estimating gender based on the expression of certain genes on chromosome Y.}

\item{model_dip}{random forest model for estimating whether chromosome arm is diploid.}

\item{model_alter}{random forest model for estimating the CNVs on chromosome arm.}

\item{chroms}{vector of chromosomes to be analyzed.}

\item{batch}{logical value, if TRUE, the samples will be normalized together as a batch, also gene expression median will be calculated from these samples}

\item{scale_cols}{colour scaling for box plots according to the median of a boxplot.}

\item{minDepth}{minimal depth of of SNV to be kept.}

\item{minReadCnt}{numeric value value used for filtering genes with low expression according to to formula: at least samp_prop*100 percent of samples have more reads than minReadCnt}

\item{samp_prop}{sample proportion which is required to have at least minReadCnt reads for a gene. The samples inlcude the diploid reference (from diploid_standard parameter) and analyzed sample.}

\item{weight_samp_prop}{proportion of samples with highest weight to be kept.}

\item{diploid_standard}{table with 50 reference diploid samples for normalizing the data.}

\item{dpRationChromEdge}{table with chromosome start and end base positions.}
}
\description{
Wrapper for generating figures and tables for CNV estimation from RNA-seq
}
