% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{vcf_to_snv}
\alias{vcf_to_snv}
\title{Fucnction that converts vcf files to tabular input for CNV analysis}
\usage{
vcf_to_snv(vcf_file, maf_tresh = 0.01, depth_tresh = 5)
}
\description{
Fucnction that converts vcf files to tabular input for CNV analysis
}
