% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{weighted_quantile}
\alias{weighted_quantile}
\title{Function for calculating weighted quantiles}
\usage{
weighted_quantile(x, w, probs = seq(0, 1, 0.25), na.rm = TRUE)
}
\arguments{
\item{x}{vector of numerical values}

\item{w}{vector of numerical weights, need to be larger than one}

\item{probs}{the quantile to be calculated}

\item{na.rm}{logical; remove NA values}
}
\description{
The code is taken from package spatstat, version 1.63.3. The function is no longer part of spatstat package
}
