% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqCNV_wrappper.R
\name{RNAseqCNV_wrapper}
\alias{RNAseqCNV_wrapper}
\title{RNAseqCNV_wrapper}
\usage{
RNAseqCNV_wrapper(
  config,
  metadata,
  adjust = TRUE,
  arm_lvl = TRUE,
  estimate_lab = TRUE,
  referData = refDataExp,
  keptSNP = keepSNP,
  par_region = par_reg,
  centr_refer = centr_ref,
  weight_tab = weight_table,
  model_gend = model_gender,
  model_dip = model_dipl,
  model_alter = model_alt,
  chroms = chrs,
  dipl_standard = diploid_standard,
  scale_cols = scaleCols,
  dpRatioChromEdge = dpRatioChrEdge,
  minDepth = 20,
  minReadCnt = 3,
  samp_prop = 0.8,
  weight_samp_prop = 1
)
}
\arguments{
\item{config}{R script assigning variables needed for the analysis}

\item{metadata}{path to a metadata table with three columns. First colum represents sample names, second file names of count files, third file names of snv files.}

\item{adjust}{logical value, determines, whether the diploid boxplots should be centered around zero on y axis}

\item{arm_lvl}{logical value, determines, wheter arm_lvl figures should be printed (increases run-time significantly)}

\item{estimate_lab}{logical value, determines, whether CNV estimation should be plotted in the final figure}

\item{referData}{table, reference data for ensamble name annotation}

\item{keptSNP}{vector of SNPs to keep for the analysis}

\item{par_region}{table with pseudoautosomal regions, in order for these regions to be fitlered out}

\item{centr_refer}{table with centromeric locations per chromosome}

\item{weight_tab}{table with per-gene weight for adjusting the importance of each gene in calling CNA}

\item{model_gend}{model for estimating gender based on expression of certain genes on chromosome Y}

\item{model_dip}{model for estimating whether chromosome arm is diploid}

\item{model_alter}{model for estimating the CNV on chromosome arm}

\item{chroms}{vector of chromosomes to be analyzed}

\item{scale_cols}{colour scaling for box plot}

\item{minDepth}{minimal depth of of SNV to be kept}

\item{minReadCnt}{numeric value value used for filtering genes with low expression according to to formula: at least samp_prop*100 percent of samples have more reads than minReadCnt}

\item{samp_prop}{sample proportion which is required to have at least minReadCnt reads for a gene}

\item{weight_samp_prop}{proportion of samples to be kept according the their weight}

\item{base_matrix}{matrix with rows being gene expression and columns DIPLOID samples. This matrix is used as a diploid reference.}

\item{base_column}{column names for base_matrix parameter}

\item{dpRationChromEdge}{table with chromosome start and end base positions}
}
\description{
Wrapper for generating figures for analysis and preview figure
}
