% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqCNVapp_wrappper.R
\name{RNAseqCNV_wrapper}
\alias{RNAseqCNV_wrapper}
\title{RNAseqCNV_wrapper}
\usage{
RNAseqCNV_wrapper(
  config,
  metadata,
  adjust = TRUE,
  arm_lvl = TRUE,
  estimate = TRUE,
  referData = refDataExp,
  keptSNP = keepSNP,
  par_region = par_reg,
  centr_refer = centr_ref,
  weight_tab = weight_table,
  model_gend = model_gender,
  model_dip = model_dipl,
  model_alter = model_alt,
  chroms = chrs,
  base_matrix = base_matr,
  base_column = base_col,
  scale_cols = scaleCols,
  dpRatioChromEdge = dpRatioChrEdge,
  minDepth = 20,
  minReadCnt = 30,
  q = 0.9,
  treshold,
  samp_perc
)
}
\arguments{
\item{config}{R script assigning variables needed for the analysis}

\item{metadata}{path to a metadata table with three columns. First colum represents sample names, second file names of count files, third file names of snv files.}

\item{adjust}{logical value, determines, whether the diploid boxplots should be centered around zero on y axis}

\item{arm_lvl}{logical value, determines, wheter arm_lvl figures should be printed (slows the computation significantly)}

\item{estimate}{logical value, determines, whether alteration will be estimated}

\item{referData}{table, reference data for ensamble name annotation}

\item{keptSNP}{vector of SNPs to keep for the analysis}

\item{par_region}{table with pseudoautosomal regions, in order for these regions to be fitlered out}

\item{centr_refer}{table with centromeric locations per chromosome}

\item{weight_tab}{table with per-gene weight for adjusting the importance of each gene in calling CNA}

\item{model_gend}{model for estimating gender based on expression of certain genes on chromosome Y}

\item{model_dip}{model for estimating whether chromosome arm is diploid}

\item{model_alter}{model for estimating the CNV on chromosome arm}

\item{chroms}{vector of chromosomes to be analyzed}

\item{base_matrix}{matrix with rows being gene expression and columns DIPLOID samples. This matrix is used as a diploid reference.}

\item{base_column}{column names for base_matrix parameter}

\item{scale_cols}{colour scaling for box plot}

\item{minDepth}{minimal depth of of SNV to be kept}

\item{minReadCnt}{numeric value value used for filtering genes with low expression according to to formula: 0.9 quantile of read count per-gene in analyzed cohort > minReadCnt}

\item{q}{numeric value, see minReadCnt}

\item{dpRationChromEdge}{table with chromosome start and end base positions}
}
\description{
Wrapper for generating figures for analysis and preview figure
}
